feat(examples): add EVSE and CEM example devices (Layer 8)
Add reference implementations demonstrating zone-based control:

EVSE (EV Charger):
- Complete device with Electrical, Measurement, EnergyControl,
  ChargingSession, and Status features
- Command handlers for SetLimit, Pause/Resume, SetChargingMode
- Simulation methods for testing: SimulateEVConnect, SimulateCharging
- Respects power limits set by CEM controller

CEM (Central Energy Manager):
- Controller that establishes zones with devices
- Methods for ConnectDevice, SetPowerLimit, ClearPowerLimit
- Subscription support for monitoring device state
- Handles notifications from subscribed devices

Also includes:
- Integration tests demonstrating zone interactions
- Fix: Use uint32 consistently for VendorID/ProductID (not uint16)

The examples show the complete MASH control flow:
1. CEM connects to EVSE
2. CEM subscribes to measurements
3. CEM sets power limits via EnergyControl commands
4. EVSE applies limits and reports effective values
5. CEM clears limits when no longer needed

+++ mash-go/pkg/features/device_info.go
@@ -198,7 +198,7 @@ func (d *DeviceInfo) SetVendorID(id uint32) error {
 }
 
 // SetProductID sets the vendor's product ID.
-func (d *DeviceInfo) SetProductID(id uint16) error {
+func (d *DeviceInfo) SetProductID(id uint32) error {
 	attr, err := d.GetAttribute(DeviceInfoAttrProductID)
 	if err != nil {
 		return err
@@ -292,9 +292,9 @@ func (d *DeviceInfo) VendorID() uint32 {
 }
 
 // ProductID returns the product ID.
-func (d *DeviceInfo) ProductID() uint16 {
+func (d *DeviceInfo) ProductID() uint32 {
 	val, _ := d.ReadAttribute(DeviceInfoAttrProductID)
-	if v, ok := val.(uint16); ok {
+	if v, ok := val.(uint32); ok {
 		return v
 	}
 	return 0
